@@ -7,12 +7,12 @@ use Symfony\Component\Console\Tester\CommandTester;
 
 class CommandFactory
 {
-    public static function tester($command, array $parameters = array(), array $options = array('decorated' => false))
+    public static function tester($command, array $parameters = [], array $options = ['decorated' => false])
     {
         $application = new Application();
         $command = $application->get($command);
         $commandTester = new CommandTester($command);
-        $input = array_merge($parameters, array('command' => $command->getName()));
+        $input = array_merge($parameters, ['command' => $command->getName()]);
         $commandTester->execute($input, $options);
 
         return $commandTester;
