@@ -23,8 +23,8 @@ class InstallerTest extends TestCase
         $this->installer = new Installer($this->filesystem, $this->zipArchive, $this->config);
 
         $this->config
-            ->shouldReceive('getOverridesSection')->andReturn(array())
-            ->shouldReceive('getOverrideFor')->andReturn(array())
+            ->shouldReceive('getOverridesSection')->andReturn([])
+            ->shouldReceive('getOverrideFor')->andReturn([])
             ->shouldReceive('getBasePackagesUrl')->andReturn('http://bower.herokuapp.com/packages/')
             ->shouldReceive('getInstallDir')->andReturn(getcwd() . '/bower_components')
             ->shouldReceive('getCacheDir')->andReturn('.')
@@ -37,7 +37,7 @@ class InstallerTest extends TestCase
 
         $package
             ->shouldReceive('getName')->andReturn('jquery')
-            ->shouldReceive('getInfo')->andReturn(array('name' => 'jquery', 'version' => '2.0.3'))
+            ->shouldReceive('getInfo')->andReturn(['name' => 'jquery', 'version' => '2.0.3'])
             ->shouldReceive('getVersion')->andReturn('2.0.3')
             ->shouldReceive('getRequiredVersion')->andReturn('2.0.3')
         ;
@@ -46,7 +46,7 @@ class InstallerTest extends TestCase
             ->shouldReceive('open')->with('./tmp/jquery')->andReturn(true)
             ->shouldReceive('getNumFiles')->andReturn(1)
             ->shouldReceive('getNameIndex')->with(0)->andReturn('jquery')
-            ->shouldReceive('statIndex')->andReturn(array('name' => 'jquery/foo', 'size' => 10, 'mtime' => 1396303200))
+            ->shouldReceive('statIndex')->andReturn(['name' => 'jquery/foo', 'size' => 10, 'mtime' => 1396303200])
             ->shouldReceive('getStream')->with('jquery/foo')->andReturn('foo content')
             ->shouldReceive('close')
         ;
@@ -71,7 +71,7 @@ class InstallerTest extends TestCase
 
         $package
             ->shouldReceive('getName')->andReturn('jquery')
-            ->shouldReceive('getInfo')->andReturn(array('name' => 'jquery', 'version' => '2.0.3'))
+            ->shouldReceive('getInfo')->andReturn(['name' => 'jquery', 'version' => '2.0.3'])
             ->shouldReceive('getRequiredVersion')->andReturn('2.0.3')
             ->shouldReceive('getVersion')->andReturn('2.0.3')
         ;
@@ -80,7 +80,7 @@ class InstallerTest extends TestCase
             ->shouldReceive('open')->with('./tmp/jquery')->andReturn(true)
             ->shouldReceive('getNumFiles')->andReturn(1)
             ->shouldReceive('getNameIndex')->with(0)->andReturn('')
-            ->shouldReceive('statIndex')->andReturn(array('name' => 'jquery/foo', 'size' => 10, 'mtime' => 1396303200))
+            ->shouldReceive('statIndex')->andReturn(['name' => 'jquery/foo', 'size' => 10, 'mtime' => 1396303200])
             ->shouldReceive('getStream')->with('jquery/foo')->andReturn('foo content')
             ->shouldReceive('close')
         ;
@@ -144,28 +144,28 @@ class InstallerTest extends TestCase
             ->shouldReceive('getNameIndex')->with(0)->andReturn('dir/')
             ->shouldReceive('getNumFiles')->andReturn(12)
             ->shouldReceive('statIndex')->times(12)->andReturn(
-                array('name' => 'dir/.foo', 'size' => 10),
-                array('name' => 'dir/foo', 'size' => 10),
-                array('name' => 'dir/foo.ext', 'size' => 12),
-                array('name' => 'dir/bar/foo.ext', 'size' => 13),
-                array('name' => 'dir/bar/anotherdir', 'size' => 13),
-                array('name' => 'dir/anotherdir', 'size' => 13),
-                array('name' => 'dir/_foo', 'size' => 3),
-                array('name' => 'dir/_fooz/bar', 'size' => 3),
-                array('name' => 'dir/filename', 'size' => 3),
-                array('name' => 'dir/okfile', 'size' => 3),
-                array('name' => 'dir/subdir/file', 'size' => 3),
-                array('name' => 'dir/zzdir/subdir/file', 'size' => 3)
+                ['name' => 'dir/.foo', 'size' => 10],
+                ['name' => 'dir/foo', 'size' => 10],
+                ['name' => 'dir/foo.ext', 'size' => 12],
+                ['name' => 'dir/bar/foo.ext', 'size' => 13],
+                ['name' => 'dir/bar/anotherdir', 'size' => 13],
+                ['name' => 'dir/anotherdir', 'size' => 13],
+                ['name' => 'dir/_foo', 'size' => 3],
+                ['name' => 'dir/_fooz/bar', 'size' => 3],
+                ['name' => 'dir/filename', 'size' => 3],
+                ['name' => 'dir/okfile', 'size' => 3],
+                ['name' => 'dir/subdir/file', 'size' => 3],
+                ['name' => 'dir/zzdir/subdir/file', 'size' => 3]
             );
         $filterZipFiles = $this->getMethod('Bowerphp\Installer\Installer', 'filterZipFiles');
-        $ignore = array('**/.*', '_*', 'subdir', '/anotherdir', 'filename',  '*.ext');
-        $expect = array(
+        $ignore = ['**/.*', '_*', 'subdir', '/anotherdir', 'filename',  '*.ext'];
+        $expect = [
             'dir/foo',
             'dir/bar/anotherdir',
             'dir/okfile',
             'dir/zzdir/subdir/file',
-        );
-        $this->assertEquals($expect, $filterZipFiles->invokeArgs($this->installer, array($archive, $ignore)));
+        ];
+        $this->assertEquals($expect, $filterZipFiles->invokeArgs($this->installer, [$archive, $ignore]));
     }
 
     public function testUninstall()
@@ -189,7 +189,7 @@ class InstallerTest extends TestCase
         $finder = Mockery::mock('Symfony\Component\Finder\Finder');
 
         $finder
-            ->shouldReceive('directories->in')->andReturn(array('package1', 'package2'));
+            ->shouldReceive('directories->in')->andReturn(['package1', 'package2']);
 
         $this->filesystem
             ->shouldReceive('exists')->with(getcwd() . '/bower_components')->andReturn(true)
@@ -210,7 +210,7 @@ class InstallerTest extends TestCase
             ->shouldReceive('exists')->with(getcwd() . '/bower_components')->andReturn(false)
         ;
 
-        $this->assertEquals(array(), $this->installer->getInstalled($finder));
+        $this->assertEquals([], $this->installer->getInstalled($finder));
     }
 
     /**
@@ -222,7 +222,7 @@ class InstallerTest extends TestCase
         $finder = Mockery::mock('Symfony\Component\Finder\Finder');
 
         $finder
-            ->shouldReceive('directories->in')->andReturn(array('package1'));
+            ->shouldReceive('directories->in')->andReturn(['package1']);
 
         $this->filesystem
             ->shouldReceive('exists')->with(getcwd() . '/bower_components')->andReturn(true)
@@ -243,7 +243,7 @@ class InstallerTest extends TestCase
         ;
 
         $finder
-            ->shouldReceive('directories->in')->andReturn(array('package1', 'package2'));
+            ->shouldReceive('directories->in')->andReturn(['package1', 'package2']);
 
         $this->filesystem
             ->shouldReceive('exists')->with(getcwd() . '/bower_components')->andReturn(true)
@@ -265,7 +265,7 @@ class InstallerTest extends TestCase
      */
     public function testIsIgnored($filename)
     {
-        $ignore = array(
+        $ignore = [
             '**/.*',
             '_*',
             'docs-assets',
@@ -280,9 +280,9 @@ class InstallerTest extends TestCase
             'composer.json',
             'package.json',
             '*.html',
-        );
+        ];
 
-        $ignored = $this->installer->isIgnored($filename, $ignore, array(), 'twbs-bootstrap-6d03173/');
+        $ignored = $this->installer->isIgnored($filename, $ignore, [], 'twbs-bootstrap-6d03173/');
         $this->assertTrue($ignored);
     }
 
@@ -291,7 +291,7 @@ class InstallerTest extends TestCase
      */
     public function testIsNotIgnored($filename)
     {
-        $ignore = array(
+        $ignore = [
             '**/.*',
             '_*',
             'docs-assets',
@@ -306,10 +306,10 @@ class InstallerTest extends TestCase
             'composer.json',
             'package.json',
             '*.html',
-        );
-        $force = array(
+        ];
+        $force = [
             'bower.json',
-        );
+        ];
 
         $ignored = $this->installer->isIgnored($filename, $ignore, $force, 'twbs-bootstrap-6d03173/');
         $this->assertFalse($ignored);
@@ -317,203 +317,203 @@ class InstallerTest extends TestCase
 
     public function providerIgnored()
     {
-        return array(
-            array('twbs-bootstrap-6d03173/.editorconfig'),
-            array('twbs-bootstrap-6d03173/.gitattributes'),
-            array('twbs-bootstrap-6d03173/.gitignore'),
-            array('twbs-bootstrap-6d03173/.travis.yml'),
-            array('twbs-bootstrap-6d03173/CNAME'),
-            array('twbs-bootstrap-6d03173/CONTRIBUTING.md'),
-            array('twbs-bootstrap-6d03173/Gruntfile.js'),
-            array('twbs-bootstrap-6d03173/_config.yml'),
-            array('twbs-bootstrap-6d03173/_includes/ads.html'),
-            array('twbs-bootstrap-6d03173/_includes/footer.html'),
-            array('twbs-bootstrap-6d03173/_includes/header.html'),
-            array('twbs-bootstrap-6d03173/_includes/nav-about.html'),
-            array('twbs-bootstrap-6d03173/_includes/nav-components.html'),
-            array('twbs-bootstrap-6d03173/_includes/nav-css.html'),
-            array('twbs-bootstrap-6d03173/_includes/nav-customize.html'),
-            array('twbs-bootstrap-6d03173/_includes/nav-getting-started.html'),
-            array('twbs-bootstrap-6d03173/_includes/nav-javascript.html'),
-            array('twbs-bootstrap-6d03173/_includes/nav-main.html'),
-            array('twbs-bootstrap-6d03173/_includes/old-bs-docs.html'),
-            array('twbs-bootstrap-6d03173/_includes/social-buttons.html'),
-            array('twbs-bootstrap-6d03173/_layouts/default.html'),
-            array('twbs-bootstrap-6d03173/_layouts/home.html'),
-            array('twbs-bootstrap-6d03173/about.html'),
-            array('twbs-bootstrap-6d03173/components.html'),
-            array('twbs-bootstrap-6d03173/composer.json'),
-            array('twbs-bootstrap-6d03173/css.html'),
-            array('twbs-bootstrap-6d03173/customize.html'),
-            array('twbs-bootstrap-6d03173/docs-assets/css/docs.css'),
-            array('twbs-bootstrap-6d03173/docs-assets/css/pygments-manni.css'),
-            array('twbs-bootstrap-6d03173/docs-assets/ico/apple-touch-icon-144-precomposed.png'),
-            array('twbs-bootstrap-6d03173/docs-assets/ico/favicon.png'),
-            array('twbs-bootstrap-6d03173/docs-assets/js/application.js'),
-            array('twbs-bootstrap-6d03173/docs-assets/js/customizer.js'),
-            array('twbs-bootstrap-6d03173/docs-assets/js/filesaver.js'),
-            array('twbs-bootstrap-6d03173/docs-assets/js/holder.js'),
-            array('twbs-bootstrap-6d03173/docs-assets/js/ie8-responsive-file-warning.js'),
-            array('twbs-bootstrap-6d03173/docs-assets/js/jszip.js'),
-            array('twbs-bootstrap-6d03173/docs-assets/js/less.js'),
-            array('twbs-bootstrap-6d03173/docs-assets/js/raw-files.js'),
-            array('twbs-bootstrap-6d03173/docs-assets/js/uglify.js'),
-            array('twbs-bootstrap-6d03173/examples/carousel/carousel.css'),
-            array('twbs-bootstrap-6d03173/examples/carousel/index.html'),
-            array('twbs-bootstrap-6d03173/examples/grid/grid.css'),
-            array('twbs-bootstrap-6d03173/examples/grid/index.html'),
-            array('twbs-bootstrap-6d03173/examples/jumbotron-narrow/index.html'),
-            array('twbs-bootstrap-6d03173/examples/jumbotron-narrow/jumbotron-narrow.css'),
-            array('twbs-bootstrap-6d03173/examples/jumbotron/index.html'),
-            array('twbs-bootstrap-6d03173/examples/jumbotron/jumbotron.css'),
-            array('twbs-bootstrap-6d03173/examples/justified-nav/index.html'),
-            array('twbs-bootstrap-6d03173/examples/justified-nav/justified-nav.css'),
-            array('twbs-bootstrap-6d03173/examples/navbar-fixed-top/index.html'),
-            array('twbs-bootstrap-6d03173/examples/navbar-fixed-top/navbar-fixed-top.css'),
-            array('twbs-bootstrap-6d03173/examples/navbar-static-top/index.html'),
-            array('twbs-bootstrap-6d03173/examples/navbar-static-top/navbar-static-top.css'),
-            array('twbs-bootstrap-6d03173/examples/navbar/index.html'),
-            array('twbs-bootstrap-6d03173/examples/navbar/navbar.css'),
-            array('twbs-bootstrap-6d03173/examples/non-responsive/index.html'),
-            array('twbs-bootstrap-6d03173/examples/non-responsive/non-responsive.css'),
-            array('twbs-bootstrap-6d03173/examples/offcanvas/index.html'),
-            array('twbs-bootstrap-6d03173/examples/offcanvas/offcanvas.css'),
-            array('twbs-bootstrap-6d03173/examples/offcanvas/offcanvas.js'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/carousel.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/grid.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/jumbotron-narrow.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/jumbotron.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/justified-nav.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/navbar-fixed.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/navbar-static.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/navbar.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/non-responsive.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/offcanvas.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/sign-in.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/starter-template.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/sticky-footer-navbar.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/sticky-footer.jpg'),
-            array('twbs-bootstrap-6d03173/examples/screenshots/theme.jpg'),
-            array('twbs-bootstrap-6d03173/examples/signin/index.html'),
-            array('twbs-bootstrap-6d03173/examples/signin/signin.css'),
-            array('twbs-bootstrap-6d03173/examples/starter-template/index.html'),
-            array('twbs-bootstrap-6d03173/examples/starter-template/starter-template.css'),
-            array('twbs-bootstrap-6d03173/examples/sticky-footer-navbar/index.html'),
-            array('twbs-bootstrap-6d03173/examples/sticky-footer-navbar/sticky-footer-navbar.css'),
-            array('twbs-bootstrap-6d03173/examples/sticky-footer/index.html'),
-            array('twbs-bootstrap-6d03173/examples/sticky-footer/sticky-footer.css'),
-            array('twbs-bootstrap-6d03173/examples/theme/index.html'),
-            array('twbs-bootstrap-6d03173/examples/theme/theme.css'),
-            array('twbs-bootstrap-6d03173/fonts/glyphicons-halflings-regular.eot'),
-            array('twbs-bootstrap-6d03173/fonts/glyphicons-halflings-regular.svg'),
-            array('twbs-bootstrap-6d03173/fonts/glyphicons-halflings-regular.ttf'),
-            array('twbs-bootstrap-6d03173/fonts/glyphicons-halflings-regular.woff'),
-            array('twbs-bootstrap-6d03173/getting-started.html'),
-            array('twbs-bootstrap-6d03173/index.html'),
-            array('twbs-bootstrap-6d03173/javascript.html'),
-            array('twbs-bootstrap-6d03173/js/.jshintrc'),
-            array('twbs-bootstrap-6d03173/js/tests/index.html'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/affix.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/alert.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/button.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/carousel.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/collapse.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/dropdown.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/modal.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/phantom.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/popover.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/scrollspy.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/tab.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/tooltip.js'),
-            array('twbs-bootstrap-6d03173/js/tests/unit/transition.js'),
-            array('twbs-bootstrap-6d03173/js/tests/vendor/jquery.js'),
-            array('twbs-bootstrap-6d03173/js/tests/vendor/qunit.css'),
-            array('twbs-bootstrap-6d03173/js/tests/vendor/qunit.js'),
-            array('twbs-bootstrap-6d03173/package.json'),
-        );
+        return [
+            ['twbs-bootstrap-6d03173/.editorconfig'],
+            ['twbs-bootstrap-6d03173/.gitattributes'],
+            ['twbs-bootstrap-6d03173/.gitignore'],
+            ['twbs-bootstrap-6d03173/.travis.yml'],
+            ['twbs-bootstrap-6d03173/CNAME'],
+            ['twbs-bootstrap-6d03173/CONTRIBUTING.md'],
+            ['twbs-bootstrap-6d03173/Gruntfile.js'],
+            ['twbs-bootstrap-6d03173/_config.yml'],
+            ['twbs-bootstrap-6d03173/_includes/ads.html'],
+            ['twbs-bootstrap-6d03173/_includes/footer.html'],
+            ['twbs-bootstrap-6d03173/_includes/header.html'],
+            ['twbs-bootstrap-6d03173/_includes/nav-about.html'],
+            ['twbs-bootstrap-6d03173/_includes/nav-components.html'],
+            ['twbs-bootstrap-6d03173/_includes/nav-css.html'],
+            ['twbs-bootstrap-6d03173/_includes/nav-customize.html'],
+            ['twbs-bootstrap-6d03173/_includes/nav-getting-started.html'],
+            ['twbs-bootstrap-6d03173/_includes/nav-javascript.html'],
+            ['twbs-bootstrap-6d03173/_includes/nav-main.html'],
+            ['twbs-bootstrap-6d03173/_includes/old-bs-docs.html'],
+            ['twbs-bootstrap-6d03173/_includes/social-buttons.html'],
+            ['twbs-bootstrap-6d03173/_layouts/default.html'],
+            ['twbs-bootstrap-6d03173/_layouts/home.html'],
+            ['twbs-bootstrap-6d03173/about.html'],
+            ['twbs-bootstrap-6d03173/components.html'],
+            ['twbs-bootstrap-6d03173/composer.json'],
+            ['twbs-bootstrap-6d03173/css.html'],
+            ['twbs-bootstrap-6d03173/customize.html'],
+            ['twbs-bootstrap-6d03173/docs-assets/css/docs.css'],
+            ['twbs-bootstrap-6d03173/docs-assets/css/pygments-manni.css'],
+            ['twbs-bootstrap-6d03173/docs-assets/ico/apple-touch-icon-144-precomposed.png'],
+            ['twbs-bootstrap-6d03173/docs-assets/ico/favicon.png'],
+            ['twbs-bootstrap-6d03173/docs-assets/js/application.js'],
+            ['twbs-bootstrap-6d03173/docs-assets/js/customizer.js'],
+            ['twbs-bootstrap-6d03173/docs-assets/js/filesaver.js'],
+            ['twbs-bootstrap-6d03173/docs-assets/js/holder.js'],
+            ['twbs-bootstrap-6d03173/docs-assets/js/ie8-responsive-file-warning.js'],
+            ['twbs-bootstrap-6d03173/docs-assets/js/jszip.js'],
+            ['twbs-bootstrap-6d03173/docs-assets/js/less.js'],
+            ['twbs-bootstrap-6d03173/docs-assets/js/raw-files.js'],
+            ['twbs-bootstrap-6d03173/docs-assets/js/uglify.js'],
+            ['twbs-bootstrap-6d03173/examples/carousel/carousel.css'],
+            ['twbs-bootstrap-6d03173/examples/carousel/index.html'],
+            ['twbs-bootstrap-6d03173/examples/grid/grid.css'],
+            ['twbs-bootstrap-6d03173/examples/grid/index.html'],
+            ['twbs-bootstrap-6d03173/examples/jumbotron-narrow/index.html'],
+            ['twbs-bootstrap-6d03173/examples/jumbotron-narrow/jumbotron-narrow.css'],
+            ['twbs-bootstrap-6d03173/examples/jumbotron/index.html'],
+            ['twbs-bootstrap-6d03173/examples/jumbotron/jumbotron.css'],
+            ['twbs-bootstrap-6d03173/examples/justified-nav/index.html'],
+            ['twbs-bootstrap-6d03173/examples/justified-nav/justified-nav.css'],
+            ['twbs-bootstrap-6d03173/examples/navbar-fixed-top/index.html'],
+            ['twbs-bootstrap-6d03173/examples/navbar-fixed-top/navbar-fixed-top.css'],
+            ['twbs-bootstrap-6d03173/examples/navbar-static-top/index.html'],
+            ['twbs-bootstrap-6d03173/examples/navbar-static-top/navbar-static-top.css'],
+            ['twbs-bootstrap-6d03173/examples/navbar/index.html'],
+            ['twbs-bootstrap-6d03173/examples/navbar/navbar.css'],
+            ['twbs-bootstrap-6d03173/examples/non-responsive/index.html'],
+            ['twbs-bootstrap-6d03173/examples/non-responsive/non-responsive.css'],
+            ['twbs-bootstrap-6d03173/examples/offcanvas/index.html'],
+            ['twbs-bootstrap-6d03173/examples/offcanvas/offcanvas.css'],
+            ['twbs-bootstrap-6d03173/examples/offcanvas/offcanvas.js'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/carousel.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/grid.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/jumbotron-narrow.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/jumbotron.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/justified-nav.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/navbar-fixed.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/navbar-static.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/navbar.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/non-responsive.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/offcanvas.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/sign-in.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/starter-template.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/sticky-footer-navbar.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/sticky-footer.jpg'],
+            ['twbs-bootstrap-6d03173/examples/screenshots/theme.jpg'],
+            ['twbs-bootstrap-6d03173/examples/signin/index.html'],
+            ['twbs-bootstrap-6d03173/examples/signin/signin.css'],
+            ['twbs-bootstrap-6d03173/examples/starter-template/index.html'],
+            ['twbs-bootstrap-6d03173/examples/starter-template/starter-template.css'],
+            ['twbs-bootstrap-6d03173/examples/sticky-footer-navbar/index.html'],
+            ['twbs-bootstrap-6d03173/examples/sticky-footer-navbar/sticky-footer-navbar.css'],
+            ['twbs-bootstrap-6d03173/examples/sticky-footer/index.html'],
+            ['twbs-bootstrap-6d03173/examples/sticky-footer/sticky-footer.css'],
+            ['twbs-bootstrap-6d03173/examples/theme/index.html'],
+            ['twbs-bootstrap-6d03173/examples/theme/theme.css'],
+            ['twbs-bootstrap-6d03173/fonts/glyphicons-halflings-regular.eot'],
+            ['twbs-bootstrap-6d03173/fonts/glyphicons-halflings-regular.svg'],
+            ['twbs-bootstrap-6d03173/fonts/glyphicons-halflings-regular.ttf'],
+            ['twbs-bootstrap-6d03173/fonts/glyphicons-halflings-regular.woff'],
+            ['twbs-bootstrap-6d03173/getting-started.html'],
+            ['twbs-bootstrap-6d03173/index.html'],
+            ['twbs-bootstrap-6d03173/javascript.html'],
+            ['twbs-bootstrap-6d03173/js/.jshintrc'],
+            ['twbs-bootstrap-6d03173/js/tests/index.html'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/affix.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/alert.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/button.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/carousel.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/collapse.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/dropdown.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/modal.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/phantom.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/popover.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/scrollspy.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/tab.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/tooltip.js'],
+            ['twbs-bootstrap-6d03173/js/tests/unit/transition.js'],
+            ['twbs-bootstrap-6d03173/js/tests/vendor/jquery.js'],
+            ['twbs-bootstrap-6d03173/js/tests/vendor/qunit.css'],
+            ['twbs-bootstrap-6d03173/js/tests/vendor/qunit.js'],
+            ['twbs-bootstrap-6d03173/package.json'],
+        ];
     }
 
     public function providerNotIgnored()
     {
-        return array(
-            array('twbs-bootstrap-6d03173/DOCS-LICENSE'),
-            array('twbs-bootstrap-6d03173/LICENSE'),
-            array('twbs-bootstrap-6d03173/LICENSE-MIT'),
-            array('twbs-bootstrap-6d03173/README.md'),
-            array('twbs-bootstrap-6d03173/bower.json'),
-            array('twbs-bootstrap-6d03173/dist/css/bootstrap-theme.css'),
-            array('twbs-bootstrap-6d03173/dist/css/bootstrap-theme.min.css'),
-            array('twbs-bootstrap-6d03173/dist/css/bootstrap.css'),
-            array('twbs-bootstrap-6d03173/dist/css/bootstrap.min.css'),
-            array('twbs-bootstrap-6d03173/dist/fonts/glyphicons-halflings-regular.eot'),
-            array('twbs-bootstrap-6d03173/dist/fonts/glyphicons-halflings-regular.svg'),
-            array('twbs-bootstrap-6d03173/dist/fonts/glyphicons-halflings-regular.ttf'),
-            array('twbs-bootstrap-6d03173/dist/fonts/glyphicons-halflings-regular.woff'),
-            array('twbs-bootstrap-6d03173/dist/js/bootstrap.js'),
-            array('twbs-bootstrap-6d03173/dist/js/bootstrap.min.js'),
-            array('twbs-bootstrap-6d03173/js/affix.js'),
-            array('twbs-bootstrap-6d03173/js/alert.js'),
-            array('twbs-bootstrap-6d03173/js/button.js'),
-            array('twbs-bootstrap-6d03173/js/carousel.js'),
-            array('twbs-bootstrap-6d03173/js/collapse.js'),
-            array('twbs-bootstrap-6d03173/js/dropdown.js'),
-            array('twbs-bootstrap-6d03173/js/modal.js'),
-            array('twbs-bootstrap-6d03173/js/popover.js'),
-            array('twbs-bootstrap-6d03173/js/scrollspy.js'),
-            array('twbs-bootstrap-6d03173/js/tab.js'),
-            array('twbs-bootstrap-6d03173/js/tooltip.js'),
-            array('twbs-bootstrap-6d03173/js/transition.js'),
-            array('twbs-bootstrap-6d03173/less/alerts.less'),
-            array('twbs-bootstrap-6d03173/less/badges.less'),
-            array('twbs-bootstrap-6d03173/less/bootstrap.less'),
-            array('twbs-bootstrap-6d03173/less/breadcrumbs.less'),
-            array('twbs-bootstrap-6d03173/less/button-groups.less'),
-            array('twbs-bootstrap-6d03173/less/buttons.less'),
-            array('twbs-bootstrap-6d03173/less/carousel.less'),
-            array('twbs-bootstrap-6d03173/less/close.less'),
-            array('twbs-bootstrap-6d03173/less/code.less'),
-            array('twbs-bootstrap-6d03173/less/component-animations.less'),
-            array('twbs-bootstrap-6d03173/less/dropdowns.less'),
-            array('twbs-bootstrap-6d03173/less/forms.less'),
-            array('twbs-bootstrap-6d03173/less/glyphicons.less'),
-            array('twbs-bootstrap-6d03173/less/grid.less'),
-            array('twbs-bootstrap-6d03173/less/input-groups.less'),
-            array('twbs-bootstrap-6d03173/less/jumbotron.less'),
-            array('twbs-bootstrap-6d03173/less/labels.less'),
-            array('twbs-bootstrap-6d03173/less/list-group.less'),
-            array('twbs-bootstrap-6d03173/less/media.less'),
-            array('twbs-bootstrap-6d03173/less/mixins.less'),
-            array('twbs-bootstrap-6d03173/less/modals.less'),
-            array('twbs-bootstrap-6d03173/less/navbar.less'),
-            array('twbs-bootstrap-6d03173/less/navs.less'),
-            array('twbs-bootstrap-6d03173/less/normalize.less'),
-            array('twbs-bootstrap-6d03173/less/pager.less'),
-            array('twbs-bootstrap-6d03173/less/pagination.less'),
-            array('twbs-bootstrap-6d03173/less/panels.less'),
-            array('twbs-bootstrap-6d03173/less/popovers.less'),
-            array('twbs-bootstrap-6d03173/less/print.less'),
-            array('twbs-bootstrap-6d03173/less/progress-bars.less'),
-            array('twbs-bootstrap-6d03173/less/responsive-utilities.less'),
-            array('twbs-bootstrap-6d03173/less/scaffolding.less'),
-            array('twbs-bootstrap-6d03173/less/tables.less'),
-            array('twbs-bootstrap-6d03173/less/theme.less'),
-            array('twbs-bootstrap-6d03173/less/thumbnails.less'),
-            array('twbs-bootstrap-6d03173/less/tooltip.less'),
-            array('twbs-bootstrap-6d03173/less/type.less'),
-            array('twbs-bootstrap-6d03173/less/utilities.less'),
-            array('twbs-bootstrap-6d03173/less/variables.less'),
-            array('twbs-bootstrap-6d03173/less/wells.less'),
-        );
+        return [
+            ['twbs-bootstrap-6d03173/DOCS-LICENSE'],
+            ['twbs-bootstrap-6d03173/LICENSE'],
+            ['twbs-bootstrap-6d03173/LICENSE-MIT'],
+            ['twbs-bootstrap-6d03173/README.md'],
+            ['twbs-bootstrap-6d03173/bower.json'],
+            ['twbs-bootstrap-6d03173/dist/css/bootstrap-theme.css'],
+            ['twbs-bootstrap-6d03173/dist/css/bootstrap-theme.min.css'],
+            ['twbs-bootstrap-6d03173/dist/css/bootstrap.css'],
+            ['twbs-bootstrap-6d03173/dist/css/bootstrap.min.css'],
+            ['twbs-bootstrap-6d03173/dist/fonts/glyphicons-halflings-regular.eot'],
+            ['twbs-bootstrap-6d03173/dist/fonts/glyphicons-halflings-regular.svg'],
+            ['twbs-bootstrap-6d03173/dist/fonts/glyphicons-halflings-regular.ttf'],
+            ['twbs-bootstrap-6d03173/dist/fonts/glyphicons-halflings-regular.woff'],
+            ['twbs-bootstrap-6d03173/dist/js/bootstrap.js'],
+            ['twbs-bootstrap-6d03173/dist/js/bootstrap.min.js'],
+            ['twbs-bootstrap-6d03173/js/affix.js'],
+            ['twbs-bootstrap-6d03173/js/alert.js'],
+            ['twbs-bootstrap-6d03173/js/button.js'],
+            ['twbs-bootstrap-6d03173/js/carousel.js'],
+            ['twbs-bootstrap-6d03173/js/collapse.js'],
+            ['twbs-bootstrap-6d03173/js/dropdown.js'],
+            ['twbs-bootstrap-6d03173/js/modal.js'],
+            ['twbs-bootstrap-6d03173/js/popover.js'],
+            ['twbs-bootstrap-6d03173/js/scrollspy.js'],
+            ['twbs-bootstrap-6d03173/js/tab.js'],
+            ['twbs-bootstrap-6d03173/js/tooltip.js'],
+            ['twbs-bootstrap-6d03173/js/transition.js'],
+            ['twbs-bootstrap-6d03173/less/alerts.less'],
+            ['twbs-bootstrap-6d03173/less/badges.less'],
+            ['twbs-bootstrap-6d03173/less/bootstrap.less'],
+            ['twbs-bootstrap-6d03173/less/breadcrumbs.less'],
+            ['twbs-bootstrap-6d03173/less/button-groups.less'],
+            ['twbs-bootstrap-6d03173/less/buttons.less'],
+            ['twbs-bootstrap-6d03173/less/carousel.less'],
+            ['twbs-bootstrap-6d03173/less/close.less'],
+            ['twbs-bootstrap-6d03173/less/code.less'],
+            ['twbs-bootstrap-6d03173/less/component-animations.less'],
+            ['twbs-bootstrap-6d03173/less/dropdowns.less'],
+            ['twbs-bootstrap-6d03173/less/forms.less'],
+            ['twbs-bootstrap-6d03173/less/glyphicons.less'],
+            ['twbs-bootstrap-6d03173/less/grid.less'],
+            ['twbs-bootstrap-6d03173/less/input-groups.less'],
+            ['twbs-bootstrap-6d03173/less/jumbotron.less'],
+            ['twbs-bootstrap-6d03173/less/labels.less'],
+            ['twbs-bootstrap-6d03173/less/list-group.less'],
+            ['twbs-bootstrap-6d03173/less/media.less'],
+            ['twbs-bootstrap-6d03173/less/mixins.less'],
+            ['twbs-bootstrap-6d03173/less/modals.less'],
+            ['twbs-bootstrap-6d03173/less/navbar.less'],
+            ['twbs-bootstrap-6d03173/less/navs.less'],
+            ['twbs-bootstrap-6d03173/less/normalize.less'],
+            ['twbs-bootstrap-6d03173/less/pager.less'],
+            ['twbs-bootstrap-6d03173/less/pagination.less'],
+            ['twbs-bootstrap-6d03173/less/panels.less'],
+            ['twbs-bootstrap-6d03173/less/popovers.less'],
+            ['twbs-bootstrap-6d03173/less/print.less'],
+            ['twbs-bootstrap-6d03173/less/progress-bars.less'],
+            ['twbs-bootstrap-6d03173/less/responsive-utilities.less'],
+            ['twbs-bootstrap-6d03173/less/scaffolding.less'],
+            ['twbs-bootstrap-6d03173/less/tables.less'],
+            ['twbs-bootstrap-6d03173/less/theme.less'],
+            ['twbs-bootstrap-6d03173/less/thumbnails.less'],
+            ['twbs-bootstrap-6d03173/less/tooltip.less'],
+            ['twbs-bootstrap-6d03173/less/type.less'],
+            ['twbs-bootstrap-6d03173/less/utilities.less'],
+            ['twbs-bootstrap-6d03173/less/variables.less'],
+            ['twbs-bootstrap-6d03173/less/wells.less'],
+        ];
     }
 
     public function providerIgnoreExceptions()
     {
-        return array(
-            array('zeroclipboard-zeroclipboard-1ec7da6/test/shared/private.tests.js.html', true),
-            array('zeroclipboard-zeroclipboard-1ec7da6/src/meta/composer.json.tmpl', true),
-            array('zeroclipboard-zeroclipboard-1ec7da6/dist/ZeroClipboard.swf', false),
-            array('zeroclipboard-zeroclipboard-1ec7da6/bower.json', false),
-        );
+        return [
+            ['zeroclipboard-zeroclipboard-1ec7da6/test/shared/private.tests.js.html', true],
+            ['zeroclipboard-zeroclipboard-1ec7da6/src/meta/composer.json.tmpl', true],
+            ['zeroclipboard-zeroclipboard-1ec7da6/dist/ZeroClipboard.swf', false],
+            ['zeroclipboard-zeroclipboard-1ec7da6/bower.json', false],
+        ];
     }
 
     /**
@@ -521,22 +521,22 @@ class InstallerTest extends TestCase
      */
     public function testIgnoreExceptions($file, $shouldIgnore)
     {
-        $ignore = array(
+        $ignore = [
             '*',
             '!/bower.json',
             '!/dist/**',
-        );
-        $ignored = $this->installer->isIgnored($file, $ignore, array(), 'zeroclipboard-zeroclipboard-1ec7da6/');
+        ];
+        $ignored = $this->installer->isIgnored($file, $ignore, [], 'zeroclipboard-zeroclipboard-1ec7da6/');
         $this->assertEquals($ignored, $shouldIgnore);
     }
 
     public function providerIgnoreAllExceptNegations()
     {
-        return array(
-            array('pippo/src/foo.js', true),
-            array('pippo/dist/.htaccess', false),
-            array('pippo/bower.json', true),
-        );
+        return [
+            ['pippo/src/foo.js', true],
+            ['pippo/dist/.htaccess', false],
+            ['pippo/bower.json', true],
+        ];
     }
 
     /**
@@ -546,11 +546,11 @@ class InstallerTest extends TestCase
      */
     public function testIgnoreAllExceptNegations($file, $shouldIgnore)
     {
-        $ignore = array(
+        $ignore = [
             '**/*',
             '!/dist/**',
-        );
-        $ignored = $this->installer->isIgnored($file, $ignore, array(), 'pippo/');
+        ];
+        $ignored = $this->installer->isIgnored($file, $ignore, [], 'pippo/');
         $this->assertEquals($ignored, $shouldIgnore);
     }
 }
