@@ -57,11 +57,11 @@ EOF
      */
     private function createDefinition()
     {
-        return new InputDefinition(array(
+        return new InputDefinition([
             new InputArgument('namespace', InputArgument::OPTIONAL, 'The namespace name'),
             new InputOption('xml', null, InputOption::VALUE_NONE, 'To output list as XML'),
             new InputOption('raw', null, InputOption::VALUE_NONE, 'To output raw command list'),
             new InputOption('format', null, InputOption::VALUE_REQUIRED, 'To output list in other formats', 'txt'),
-        ));
+        ]);
     }
 }
