@@ -30,12 +30,12 @@ class HelpCommand extends SymfonyHelpCommand
 
         $this
             ->setName('help')
-            ->setDefinition(array(
+            ->setDefinition([
                 new InputArgument('command_name', InputArgument::OPTIONAL, 'The command name', 'help'),
                 new InputOption('xml', null, InputOption::VALUE_NONE, 'To output help as XML'),
                 new InputOption('format', null, InputOption::VALUE_REQUIRED, 'The output format (txt, xml, json, or md)', 'txt'),
                 new InputOption('raw', null, InputOption::VALUE_NONE, 'To output raw command help'),
-            ))
+            ])
             ->setDescription('Displays help for a command')
             ->setHelp(<<<EOF
 The <info>%command.name%</info> command displays help for a given command:
