@@ -44,7 +44,7 @@ class Application extends BaseApplication
     public function __construct()
     {
         ErrorHandler::register();
-        parent::__construct('Bowerphp', '0.3 Powered by BeeLab (bee-lab.net)');
+        parent::__construct('Bowerphp', '0.5 Powered by BeeLab (bee-lab.net)');
     }
 
     /**
@@ -52,10 +52,6 @@ class Application extends BaseApplication
      */
     public function doRun(InputInterface $input, OutputInterface $output)
     {
-        if (version_compare(PHP_VERSION, '5.3.2', '<')) {
-            $output->writeln('<warning>Bowerphp only officially supports PHP 5.3.2 and above, you will most likely encounter problems with your PHP ' . PHP_VERSION . ', upgrading is strongly recommended.</warning>');
-        }
-
         if ($input->hasParameterOption('--profile')) {
             $startTime = microtime(true);
         }
@@ -91,7 +87,7 @@ class Application extends BaseApplication
      */
     protected function getDefaultCommands()
     {
-        return array(
+        return [
             new Command\HelpCommand(),
             new Command\CommandListCommand(),
             new Command\HomeCommand(),
@@ -103,7 +99,7 @@ class Application extends BaseApplication
             new Command\SearchCommand(),
             new Command\UpdateCommand(),
             new Command\UninstallCommand(),
-        );
+        ];
     }
 
     /**
@@ -139,7 +135,7 @@ class Application extends BaseApplication
      */
     private function getNewWorkingDir(InputInterface $input)
     {
-        $workingDir = $input->getParameterOption(array('--working-dir', '-d'));
+        $workingDir = $input->getParameterOption(['--working-dir', '-d']);
         if (false !== $workingDir && !is_dir($workingDir)) {
             throw new \RuntimeException('Invalid working directory specified.');
         }
@@ -159,7 +155,7 @@ class Application extends BaseApplication
      */
     private function symfonyDoRun(InputInterface $input, OutputInterface $output, $default = 'list-commands')
     {
-        if (true === $input->hasParameterOption(array('--version', '-V'))) {
+        if (true === $input->hasParameterOption(['--version', '-V'])) {
             $output->writeln($this->getLongVersion());
 
             return 0;
@@ -167,10 +163,10 @@ class Application extends BaseApplication
 
         $name = $this->getCommandName($input);
 
-        if (true === $input->hasParameterOption(array('--help', '-h'))) {
+        if (true === $input->hasParameterOption(['--help', '-h'])) {
             if (!$name) {
                 $name = 'help';
-                $input = new ArrayInput(array('command' => 'help'));
+                $input = new ArrayInput(['command' => 'help']);
             } else {
                 $this->wantHelps = true;
             }
@@ -178,7 +174,7 @@ class Application extends BaseApplication
 
         if (!$name) {
             $name = $default;
-            $input = new ArrayInput(array('command' => $default));
+            $input = new ArrayInput(['command' => $default]);
         }
 
         // the command name MUST be the first element of the input
