@@ -13,7 +13,7 @@ $includeIfExists = function ($file) {
     return file_exists($file) ? include $file : false;
 };
 
-if ((!$loader = $includeIfExists(__DIR__ . '/../vendor/autoload.php')) && (!$loader = $includeIfExists(__DIR__ . '/../../../autoload.php'))) {
+if (!$loader = $includeIfExists(__DIR__ . '/../vendor/autoload.php')) {
     $error = 'You must set up the project dependencies, run the following commands:' . PHP_EOL .
         'curl -sS https://getcomposer.org/installer | php' . PHP_EOL .
         'php composer.phar install' . PHP_EOL;
