@@ -14,7 +14,7 @@ class InitCommandTest extends \PHPUnit_Framework_TestCase
     {
         $application = new Application();
         $commandTester = new CommandTester($command = $application->get('init'));
-        $commandTester->execute(array('command' => $command->getName()), array('interactive' => false, 'decorated' => false));
+        $commandTester->execute(['command' => $command->getName()], ['interactive' => false, 'decorated' => false]);
 
         $json = json_decode(file_get_contents(getcwd() . '/bower.json'), true);
         $this->assertArrayHasKey('name', $json);
